refactor: rename dashboard routes to patient
- Renamed all dashboard-related routes and components to patient to better reflect the user role and improve code clarity.
- Updated import paths in AppRoutes.jsx to reflect the new patient directory structure.
- Adjusted links and navigation items in DashboardLayout, Header, MobileSidebar, and Sidebar components to use the patient prefix.
