+++ src/pages/dashboard/patient/SettingsPage.jsx
@@ -1,25 +1,25 @@
-import { Bot } from "lucide-react";
-import AnalysisForm from "@/components/features/dashboard/analysis/AnalysisForm";
+import { Settings } from "lucide-react";
+import SettingsForm from "@/components/features/dashboard/patient/settings/SettingsForm";
 
-const AnalysisPage = () => {
+const SettingsPage = () => {
 	return (
 		<section className="space-y-6 py-4 sm:ml-[16rem] sm:py-8 md:space-y-8">
 			<div className="rounded-xl border border-green-200 bg-green-50 p-4 shadow sm:p-6">
 				<div className="flex items-start space-x-3 sm:space-x-4">
 					<div className="mt-1 flex-shrink-0">
-						<Bot className="h-8 w-8 text-green-600 sm:h-10 sm:w-10" />
+						<Settings className="h-8 w-8 text-green-600 sm:h-10 sm:w-10" />
 					</div>
 					<div>
-						<h2 className="text-xl font-semibold text-green-800 sm:text-2xl">Analisis Baru</h2>
+						<h2 className="text-xl font-semibold text-green-800 sm:text-2xl">Pengaturan Akun</h2>
 						<p className="mt-1 text-sm leading-relaxed text-green-700 sm:text-base">
-							Buat analisis baru dan dapatkan rekomendasi jamu yang sesuai dengan kebutuhan Anda.
+							Kelola keamanan akun Anda di sini.
 						</p>
 					</div>
 				</div>
 			</div>
-			<AnalysisForm />
+			<SettingsForm />
 		</section>
 	);
 };
 
-export default AnalysisPage;
+export default SettingsPage;
